Continuing to try to fix timezones
Sequelize (inexplicably) sets the system timezone to UTC unless you
account for it, so we will! The user will still see data that doesn’t
fit their understanding of time (if they’re in EST, this will likely
still show UTC data) but at least it’s not an extra 5 hours wrong.